Struct leetcode_rust::common::CaseGroup

pub struct CaseGroup<T, G, P> { /* private fields */ }

A easy to use test case collection struct that also provide functions for simple test case creation.

Implementations

impl<T, G, P> CaseGroup<T, G, P>

pub fn new() -> CaseGroup<T, G, P>

Create a new CaseGroup instance.

pub fn add(&mut self, case: Case<T, G, P>)

Add existing test case instance to the collection.

Note: an additional index value will be set by calling this method.

pub fn all(self) -> Vec<Case<T, G, P>>

Get all test cases within current test case collection.

impl<G, P> CaseGroup<String, G, P>

Implement two handy methods on CaseGroup<String, G, P> struct.

pub fn create(&mut self, ipt: &str, exp: Vec<G>)where
P: PartialEq + Display,
G: PartialEq + Display,

Create a new test case (mo input parameters) matching &str and other generic types.

Argument

ipt - this argument is set to &str to simplify method calls.

exp - expected values in Vec<G> form.

pub fn create_param(&mut self, ipt: &str, exp: Vec<G>, params: Vec<P>)where
G: PartialEq + Display,
P: PartialEq + Display,

Create a new test case (with input parameters) matching &str and other generic types.

Argument

ipt - this argument is set to &str to simplify method calls.

exp - expected values in Vec<G> form.

params - expected values in Vec<P> form.

impl CaseGroup<i32, i32, i32>

Implement two handy methods on CaseGroup struct.

pub fn create(&mut self, ipt: i32, exp: Vec<i32>)

Create a new test case (no input parameters) matching selected generic types.

pub fn create_param(&mut self, ipt: i32, exp: Vec<i32>, params: Vec<i32>)

Create a new test case (with input parameters) matching selected generic types.