Struct leetcode_rust::common::VectorCase

pub struct VectorCase<T, G, P> {
    pub inputs: Vec<T>,
    pub params: Vec<P>,
    pub values: Vec<G>,
    /* private fields */
}

Fields
inputs: Vec<T>

Input values of test case

params: Vec<P>

Optional parameters when executing test case

values: Vec<G>

Expected values of given input

Implementations

impl<T, G, P> VectorCase<T, G, P>where
T: PartialEq,
G: PartialEq,
P: PartialEq,

pub fn new(input: Vec<T>, values: Vec<G>) -> VectorCase<T, G, P>

Create new test case with no parameters

Arguments

input - test input

values - expected values, accept single- or multi-value vector

pub fn new_multi(inputs: Vec<T>, values: Vec<G>) -> VectorCase<T, G, P>

Create new test case with no parameters but multiple inputs

Arguments

inputs - test input, accept single- or multi-value vector

values - expected values, accept single- or multi-value vector

pub fn new_params(
input: Vec<T>,
params: Vec<P>,
values: Vec<G>
) -> VectorCase<T, G, P>

Create new test case with parameters

Arguments

input - test input

params - test parameters that vary among different cases

values - expected values, accept single- or multi-value vector

pub fn new_params_multi(
inputs: Vec<T>,
params: Vec<P>,
values: Vec<G>
) -> VectorCase<T, G, P>

Create new test case with parameters and multi input

Arguments

inputs - test input, accept single- or multi-value vector

params - test parameters that vary among different cases

values - expected values, accept single- or multi-value vector

pub fn is_valid(&self, result: G)where
G: Debug,
T: Debug,

Check if solution output matches any expectations

Arguments

&self - inmutable borrow to object itself

result - solution output

use leetcode_rust::common::VectorCase;
let case:VectorCase<String, String, i32> = VectorCase::new(String::from("A"), vec![String::from("A")]);
case.is_valid(String::from("A"))

pub fn label(&self) -> String

Returns the index value in String form.

pub fn input(&self) -> &T

Returns the first element of inputs