Function leetcode_rust::problems::p000_0xx::p000_007::reverse_integer

pub fn reverse_integer(x: i32) -> i32

Reverse an integer (32-bit long) and check for overflow.

Exchange front and end digits (not bits) one by one, return zero if overflow

Description

Given a signed 32-bit integer x, return x with its digits reversed. If reversing x causes the value to go outside the signed 32-bit integer range [-2^31, 2^31 - 1], then return 0.

Assume the environment does not allow you to store 64-bit integers (signed or unsigned).

Example 1

Input: x = 123

Output: 321

Example 2

Input: x = -123

Output: -321

Example 3

Input: x = 120

Output: 21

Constraints:

-2^31 <= x <= 2^31 - 1

Source: https://leetcode.com/problems/reverse-integer/description/

Argument

x - 32-bit signed integer to alter

use leetcode_rust::problems::p000_0xx::p000_007::reverse_integer;
assert_eq!(reverse_integer(-2147483647), 0);
assert_eq!(reverse_integer(123), 321);
assert_eq!(reverse_integer(120), 21);
assert_eq!(reverse_integer(-123), -321);